requirements.txt
openai
coinbase
python-dotenv
requests
pandas
pandas_ta